requirements.txt
lxml
frogress
boto3
python-dotenv
psycopg2
urllib3